# docs/swagger.yaml
basePath: /api
definitions:
  dto.BulletinIn:
    properties:
      imageUrl:
        type: string
      title:
        type: string
    type: object
  dto.BulletinOut:
    properties:
      id:
        type: integer
      imageUrl:
        type: string
      title:
        type: string
    type: object
  dto.UserLoginIn:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  dto.UserOut:
    properties:
      avatar:
        type: string
      id:
        type: integer
      role:
        items:
          type: string
        type: array
      username:
        type: string
    type: object
host: localhost
info:
  contact:
    email: t117503445@gmail.com
    name: "117503445"
    url: http://www.117503445.top
  license:
    name: GNU GPL 3.0
    url: https://github.com/TGclub/smartlab_backend_go/blob/main/LICENSE
  title: smartlab
  version: "1.0"
paths:
  /Bulletin:
    get:
      consumes:
      - application/json
      description: 读取所有公告
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.BulletinOut'
            type: array
      summary: BulletinReadAll
    post:
      consumes:
      - application/json
      description: 创建公告，需要管理员权限。
      parameters:
      - description: dto.BulletinIn
        in: body
        name: BulletinIn
        required: true
        schema:
          $ref: '#/definitions/dto.BulletinIn'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.BulletinOut'
            type: array
      security:
      - JWT: []
      summary: BulletinCreate
  /Bulletin/{id}:
    delete:
      consumes:
      - application/json
      description: 删除公告，需要管理员权限。
      parameters:
      - description: DeleteBulletin.ID
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.BulletinOut'
            type: array
      security:
      - JWT: []
      summary: BulletinDelete
    get:
      consumes:
      - application/json
      description: 读取指定公告
      parameters:
      - description: Bulletin.ID
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.BulletinOut'
            type: array
      summary: BulletinRead
  /ping:
    post:
      consumes:
      - application/json
      description: 返回 pong
      operationId: ping
      produces:
      - application/json
      responses:
        "200":
          description: pong
          schema:
            type: string
      summary: 状态检查
  /user/login:
    post:
      consumes:
      - application/json
      description: 登陆账户，返回 JWT
      parameters:
      - description: dto.userLoginIn
        in: body
        name: userLoginIn
        required: true
        schema:
          $ref: '#/definitions/dto.UserLoginIn'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.UserOut'
            type: array
      summary: Login
schemes:
- http
securityDefinitions:
  JWT:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
x-extension-openapi:
  example: value on a json format